+++ chef-repo/site-cookbooks/millefeuille/CHANGELOG.md
@@ -1,9 +1,9 @@
-# hello CHANGELOG
+# millefeuille CHANGELOG
 
-This file is used to list changes made in each version of the hello cookbook.
+This file is used to list changes made in each version of the millefeuille cookbook.
 
 ## 0.1.0
-- [your_name] - Initial release of hello
+- [your_name] - Initial release of millefeuille
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.